Move GradientController to top-level components
- Remove controller interface and related files
- Migrate GradientController to be a top-level component
